% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causens.R
\name{causens}
\alias{causens}
\title{Causal sensitivity analysis}
\usage{
causens(out.model, trt.model, data, exposure, outcome, sf)
}
\arguments{
\item{out.model}{The outcome model object on which we wish to perform a sensitivity analysis}

\item{trt.model}{The treatment model object}
}
\value{
NA
}
\description{
TODO: Write description
}
