% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causens_monte_carlo.R
\name{causens_monte_carlo}
\alias{causens_monte_carlo}
\title{Monte Carlo sensitivity analysis for causal effects}
\usage{
causens_monte_carlo(outcome, exposure, confounders, data, ...)
}
\arguments{
\item{outcome}{The name of the outcome variable in the data frame.}

\item{exposure}{The name of the exposure variable in the data frame.}

\item{confounders}{The name of the confounders in the data frame.}

\item{data}{A data frame containing the exposure, outcome, and confounder
variables.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
The estimated causal effect.
}
\description{
This function performs a Monte Carlo sensitivity analysis for
causal effects.
}
