% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_ate.R
\name{corrected_ate}
\alias{corrected_ate}
\title{Corrected ATE}
\usage{
corrected_ate(trt_model, data, exposure, outcome, sf)
}
\arguments{
\item{trt_model}{The treatment model object.}

\item{data}{A data frame containing the variables of interest.}

\item{exposure}{The name of the exposure variable.}

\item{outcome}{The name of the outcome variable.}

\item{sf}{The R sensitivity function for c(z, e).}
}
\value{
A vector corresponding with the corrected ATE.
}
\description{
This function provides the corrected ATE using a sensitivity
function c(z, e).
}
\examples{
data <- data.frame(
  exposure = c(0, 1, 0, 1),
  outcome = c(0, 1, 0, 1),
  confounder1 = c(1, 0, 1, 0),
  confounder2 = c(0, 1, 0, 1)
)
corrected_ate(data, "exposure", "outcome", c(
  "confounder1",
  "confounder2"
), 0.5)
}
