% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_corrected_outcomes.R
\name{corrected_outcomes}
\alias{corrected_outcomes}
\title{Outcomes Corrected via Sensitivity Function}
\usage{
corrected_outcomes(
  trt_model,
  data,
  exposure,
  outcome,
  sf = "constant",
  c1 = 0.5,
  c0 = 0.3
)
}
\arguments{
\item{trt_model}{The treatment model object.}

\item{data}{A data frame containing the variables of interest.}

\item{exposure}{The name of the exposure variable.}

\item{outcome}{The name of the outcome variable.}

\item{sf}{The R sensitivity function for c(z, e).}

\item{confounders}{A vector of names of confounding variables.}
}
\value{
A vector corresponding with corrected outcomes.
}
\description{
This function provides corrected outcomes corrected using a sensitivity
function c(z, e).
}
\examples{
data <- data.frame(
  exposure = c(0, 1, 0, 1),
  outcome = c(0, 1, 0, 1),
  confounder1 = c(1, 0, 1, 0),
  confounder2 = c(0, 1, 0, 1)
)
corrected_outcomes(data, "exposure", "outcome", c(
  "confounder1",
  "confounder2"
), 0.5)

}
