% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{gData_U_cont_Y_binary}
\alias{gData_U_cont_Y_binary}
\title{Generate data with a continuous unmeasured confounder and a binary outcome}
\usage{
gData_U_cont_Y_binary(
  ymodel = "linear",
  N = 500,
  alpha_uz = 0.2,
  beta_uz = 0.5,
  treatment_effects = 1,
  seed = 123
)
}
\arguments{
\item{ymodel}{A string indicating the functional form of the outcome model.}

\item{N}{The number of observations to be generated.}

\item{alpha_uz}{The coefficient of the unmeasured confounder in the propensity score model.}

\item{treatment_effects}{The treatment effect.}

\item{seed}{The seed for the random number generator.}

\item{beta_uy}{The coefficient of the unmeasured confounder in the outcome model.}
}
\description{
Generate data with a continuous unmeasured confounder and a binary outcome
}
