% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{gData_U_cont_Y_cont}
\alias{gData_U_cont_Y_cont}
\title{Generate data with a continuous unmeasured confounder and continuous
outcome}
\usage{
gData_U_cont_Y_cont(
  ymodel = "linear",
  N = 500,
  alpha_uz = 0.2,
  beta_uy = 0.5,
  treatment_effects = 1,
  seed = 123
)
}
\arguments{
\item{ymodel}{A string indicating the functional form of the outcome model.}

\item{N}{The number of observations to be generated.}

\item{alpha_uz}{Unmeasured confounder coefficient in the propensity score
model.}

\item{beta_uy}{Unmeasured confounder coefficient in the outcome model.}

\item{treatment_effects}{The treatment effect.}

\item{seed}{The seed for the random number generator.}
}
\value{
A data frame with the simulated dataset with U continuous, Y
continuous.
}
\description{
Generate data with a continuous unmeasured confounder and continuous
outcome
}
