% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_formula.R
\name{process_model_formula}
\alias{process_model_formula}
\title{Process model formula}
\usage{
process_model_formula(model, data)
}
\arguments{
\item{model}{The model object as a formula or fitted glm.}

\item{data}{A data frame used for fitting the model.}
}
\value{
A list containing the fitted model, model formula, response variable
name, and confounder names.
}
\description{
This helper function takes in a model in its `formula` format
and returns a list comprising of various information, including the fitted
model as an R object.
}
