% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.causens_sf}
\alias{summary.causens_sf}
\title{Summarize the results of a causal sensitivity analysis via
sensitivity function.}
\usage{
\method{summary}{causens_sf}(causens_obj)
}
\arguments{
\item{causens_obj}{An object of class \code{causens_sf}.}
}
\value{
A summary of the results of the causal sensitivity analysis.
}
\description{
Summarize the results of a causal sensitivity analysis via
sensitivity function.
}
