% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.monte_carlo_causens}
\alias{summary.monte_carlo_causens}
\title{Summarize the results of a causal sensitivity analysis via the
Monte Carlo method.}
\usage{
\method{summary}{monte_carlo_causens}(object, ...)
}
\arguments{
\item{object}{An object of class \code{causens_monte_carlo}.}

\item{...}{Additional arguments to be passed to \code{summary}.}
}
\value{
A summary of results from a Monte Carlo causal sensitivity analysis.
}
\description{
Summarize the results of a causal sensitivity analysis via the
Monte Carlo method.
}
