% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_causens.R
\name{bayesian_causens}
\alias{bayesian_causens}
\title{Bayesian parametric sensitivity analysis for causal inference}
\usage{
bayesian_causens(
  exposure,
  outcome,
  confounders,
  data,
  backend = "jags",
  output_trace = FALSE,
  ...
)
}
\arguments{
\item{exposure}{Exposure variable name in the data frame.}

\item{outcome}{Outcome variable name in the data frame.}

\item{confounders}{Confounders' names in the data frame.}

\item{data}{A data frame containing the exposure, outcome, and confounder variables.}

\item{backend}{The backend to use for the sensitivity analysis. Currently
only "jags" is supported.}

\item{output_trace}{Whether to output the full trace of the MCMC sampler.}

\item{...}{Additional arguments to be passed to the backend.}
}
\value{
A list of posterior samples for the causal effect of the exposure
variable on the outcome, as well as the confounder-adjusted causal effect.
}
\description{
This function runs a Bayesian sensitivity analysis for causal
inference using JAGS or Stan as a backend.
}
