% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_causens.R
\name{bayesian_causens}
\alias{bayesian_causens}
\title{Bayesian parametric sensitivity analysis for causal inference}
\usage{
bayesian_causens(data, exposure, outcome, confounders, backend = "jags", ...)
}
\arguments{
\item{data}{A data frame containing the exposure, outcome, and confounders.}

\item{exposure}{The name of the exposure variable in the data frame.}

\item{outcome}{The name of the outcome variable in the data frame.}

\item{confounders}{The name of the confounders in the data frame.}

\item{backend}{The backend to use for the sensitivity analysis. Currently
only "jags" is supported.}

\item{...}{Additional arguments to be passed to the backend.}
}
\value{
A list of posterior samples for the causal effect of the exposure
variable on the outcome, as well as the confounder-adjusted causal effect.
}
\description{
This function runs a Bayesian sensitivity analysis for causal
inference using JAGS or Stan as a backend.
}
\details{
Tasks
1. correct parameter naming
2. make sure it doesn't crash
3. run on simulated data
}
