% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causens.R
\name{causens}
\alias{causens}
\title{Causal Effect Estimation with Sensitivity Analysis}
\usage{
causens(trt_model, outcome, method, data, ...)
}
\arguments{
\item{trt_model}{The treatment model object as a formula or fitted glm.}

\item{outcome}{The name of the outcome variable.}

\item{method}{The method to use for sensitivity analysis. Currently, "Li" and
"Bayesian" are supported.}

\item{data}{A data frame containing the exposure, outcome, and confounder variables.}

\item{...}{Additional arguments to be passed to the sensitivity function.}
}
\value{
A point estimate of the corrected ATE.
}
\description{
This function provides an estimate of the Average Treatment
Effect (ATE) using outcomes corrected for unmeasured confounding via a
sensitivity function.

Arguments in the ellipsis are 'keyword arguments' and are passed
to the sensitivity function `sf`.
}
