% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causens.R
\name{causens}
\alias{causens}
\title{Estimate Corrected Average Treatment Effect}
\usage{
causens(trt_model, data, exposure, outcome, c1 = 0.5, c0 = 0.3, s1 = 0, s0 = 0)
}
\arguments{
\item{trt_model}{The treatment model object.}

\item{data}{A data frame containing the variables of interest.}

\item{exposure}{The name of the exposure variable.}

\item{outcome}{The name of the outcome variable.}

\item{confounders}{A vector of names of confounding variables.}

\item{sf}{The R sensitivity function for c(z, e).}
}
\value{
A point estimate of the corrected ATE.
}
\description{
This function provides an estimate of the Average Treatment Effect (ATE) using
outcomes corrected for unmeasured confounding via a sensitivity function.
}
