% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causens.R
\name{causens}
\alias{causens}
\title{Causal Effect Estimation with Sensitivity Analysis}
\usage{
causens(trt_model, data, exposure, outcome, method, ...)
}
\arguments{
\item{trt_model}{The treatment model object.}

\item{data}{A data frame containing the variables of interest.}

\item{exposure}{The name of the exposure variable.}

\item{outcome}{The name of the outcome variable.}

\item{method}{The method to use for sensitivity analysis. Currently, only
"Li" is supported.}

\item{confounders}{A vector of names of confounding variables.}
}
\value{
A point estimate of the corrected ATE.
}
\description{
This function provides an estimate of the Average Treatment
Effect (ATE) using outcomes corrected for unmeasured confounding via a
sensitivity function.

Arguments in the ellipsis are 'keyword arguments' and are passed
to the sensitivity function `sf`.
}
