% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causens_sf.R
\name{causens_sf}
\alias{causens_sf}
\title{Bayesian Estimation of ATE Subject to Unmeasured Confounding}
\usage{
causens_sf(fitted_model, exposure, outcome, data, ...)
}
\arguments{
\item{fitted_model}{The treatment model object as a glm.}

\item{exposure}{The name of the exposure variable.}

\item{outcome}{The name of the outcome variable.}

\item{data}{A data frame containing the exposure, outcome, and confounder variables.}

\item{...}{Additional arguments to be passed to the sensitivity function.}
}
\value{
A point estimate of the corrected ATE.
}
\description{
This function provides an estimate of the Average Treatment
Effect (ATE) using Bayesian modelling.
}
