% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bayesian_causens}
\alias{summary.bayesian_causens}
\title{Summarize the results of a causal sensitivity analysis via
Bayesian modelling of an unmeasured confounder.}
\usage{
\method{summary}{bayesian_causens}(causens_obj)
}
\arguments{
\item{causens_obj}{An object of class \code{causens_bayesian}.}
}
\description{
Summarize the results of a causal sensitivity analysis via
Bayesian modelling of an unmeasured confounder.
}
