% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bayesian_causens}
\alias{summary.bayesian_causens}
\title{Summarize the results of a causal sensitivity analysis via
Bayesian modelling of an unmeasured confounder.}
\usage{
\method{summary}{bayesian_causens}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bayesian_causens}.}

\item{...}{Additional arguments to be passed to \code{summary}.}
}
\value{
A summary of results from a Bayesian causal sensitivity analysis.
}
\description{
Summarize the results of a causal sensitivity analysis via
Bayesian modelling of an unmeasured confounder.
}
